pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["AlexGrig23 <alexgrig.cyber@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
requests = "^2.31.0"
instaloader = "^4.10.2"


[tool.poetry.group.dev.dependencies]
django-admin = "^2.0.2"
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
